requirements.txt
streamlit==1.13.0
numpy==1.23.5
opencv-python-headless==4.7.0.72
Pillow==8.2.0
matplotlib==3.3.4
scipy==1.10.1
mtcnn==0.1.1
keras-vggface==0.6